requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
autobahn==19.11.2
Automat==25.4.16
beautifulsoup4==4.9.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==7.1.2
constantly==23.10.4
cryptography==46.0.5
curl_cffi==0.13.0
distro==1.9.0
frozendict==2.4.7
google-auth==2.49.0
google-genai==1.66.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jugaad-data==0.29
kiteconnect==5.0.1
lxml==6.0.2
markdown==3.5.2
multitasking==0.0.12
nsepy==0.8
nsepython==2.97
numpy==2.4.2
packaging==26.0
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.4
protobuf==7.34.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
rsa==4.9.1
scipy==1.17.1
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
ta==0.11.0
tenacity==9.1.4
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
zope.interface
pyotp>=2.9.0
